+++ sentry-ruby/lib/sentry/rack/capture_exceptions.rb
@@ -18,7 +18,7 @@ module Sentry
         Sentry.with_scope do |scope|
           Sentry.with_session_tracking do
             scope.clear_breadcrumbs
-            scope.set_transaction_name(env["PATH_INFO"]) if env["PATH_INFO"]
+            scope.set_transaction_name(env["PATH_INFO"], source: :url) if env["PATH_INFO"]
             scope.set_rack_env(env)
 
             transaction = start_transaction(env, scope)
@@ -65,7 +65,7 @@ module Sentry
         sentry_trace = env["HTTP_SENTRY_TRACE"]
         baggage = env["HTTP_BAGGAGE"]
 
-        options = { name: scope.transaction_name, op: transaction_op }
+        options = { name: scope.transaction_name, source: scope.transaction_source, op: transaction_op }
         transaction = Sentry::Transaction.from_sentry_trace(sentry_trace, baggage: baggage, **options) if sentry_trace
         Sentry.start_transaction(transaction: transaction, custom_sampling_context: { env: env }, **options)
       end