+++ sentry-ruby/lib/sentry/rake.rb
@@ -10,7 +10,7 @@ module Sentry
       def display_error_message(ex)
         Sentry.capture_exception(ex) do |scope|
           task_name = top_level_tasks.join(' ')
-          scope.set_transaction_name(task_name)
+          scope.set_transaction_name(task_name, source: :task)
           scope.set_tag("rake_task", task_name)
         end if Sentry.initialized? && !Sentry.configuration.skip_rake_integration